Module ext_websocket::actions

Structs

AddClass

AddDifferential

AddPermanentClass

AddProduct

AddProductDifferential

AddProductType

BlockRefresh
This blocks the sseq object from recomputing classes and edges. This is useful when performing a large number of operations in a row, e.g. when loading files or undoing.

Clear

Complete

Construct

ConstructJson

Message

QueryCocycleString

QueryCocycleStringResult

QueryTable

QueryTableResult

Resolve

Resolving

SetClass

SetClassName

SetDifferential

SetPageList

SetStructline

Enums

Action
This is just a list of everything that implements ActionT. We use this instead of Box<dyn ActionT> so that Serde is happy.

Recipient

SseqChoice

Constants

_IMPL_DESERIALIZE_FOR_Action

_IMPL_DESERIALIZE_FOR_AddClass

_IMPL_DESERIALIZE_FOR_AddDifferential

_IMPL_DESERIALIZE_FOR_AddPermanentClass

_IMPL_DESERIALIZE_FOR_AddProduct

_IMPL_DESERIALIZE_FOR_AddProductDifferential

_IMPL_DESERIALIZE_FOR_AddProductType

_IMPL_DESERIALIZE_FOR_BlockRefresh

_IMPL_DESERIALIZE_FOR_Clear

_IMPL_DESERIALIZE_FOR_Complete

_IMPL_DESERIALIZE_FOR_Construct

_IMPL_DESERIALIZE_FOR_ConstructJson

_IMPL_DESERIALIZE_FOR_Message

_IMPL_DESERIALIZE_FOR_QueryCocycleString

_IMPL_DESERIALIZE_FOR_QueryCocycleStringResult

_IMPL_DESERIALIZE_FOR_QueryTable

_IMPL_DESERIALIZE_FOR_QueryTableResult

_IMPL_DESERIALIZE_FOR_Recipient

_IMPL_DESERIALIZE_FOR_Resolve

_IMPL_DESERIALIZE_FOR_Resolving

_IMPL_DESERIALIZE_FOR_SetClass

_IMPL_DESERIALIZE_FOR_SetClassName

_IMPL_DESERIALIZE_FOR_SetDifferential

_IMPL_DESERIALIZE_FOR_SetPageList

_IMPL_DESERIALIZE_FOR_SetStructline

_IMPL_DESERIALIZE_FOR_SseqChoice

_IMPL_SERIALIZE_FOR_Action

_IMPL_SERIALIZE_FOR_AddClass

_IMPL_SERIALIZE_FOR_AddDifferential

_IMPL_SERIALIZE_FOR_AddPermanentClass

_IMPL_SERIALIZE_FOR_AddProduct

_IMPL_SERIALIZE_FOR_AddProductDifferential

_IMPL_SERIALIZE_FOR_AddProductType

_IMPL_SERIALIZE_FOR_BlockRefresh

_IMPL_SERIALIZE_FOR_Clear

_IMPL_SERIALIZE_FOR_Complete

_IMPL_SERIALIZE_FOR_Construct

_IMPL_SERIALIZE_FOR_ConstructJson

_IMPL_SERIALIZE_FOR_Message

_IMPL_SERIALIZE_FOR_QueryCocycleString

_IMPL_SERIALIZE_FOR_QueryCocycleStringResult

_IMPL_SERIALIZE_FOR_QueryTable

_IMPL_SERIALIZE_FOR_QueryTableResult

_IMPL_SERIALIZE_FOR_Recipient

_IMPL_SERIALIZE_FOR_Resolve

_IMPL_SERIALIZE_FOR_Resolving

_IMPL_SERIALIZE_FOR_SetClass

_IMPL_SERIALIZE_FOR_SetClassName

_IMPL_SERIALIZE_FOR_SetDifferential

_IMPL_SERIALIZE_FOR_SetPageList

_IMPL_SERIALIZE_FOR_SetStructline

_IMPL_SERIALIZE_FOR_SseqChoice

Traits

ActionT
The name Action is sort-of a misnomer. It is the content of any message that is sent between the different objects. There are three functions to implement.