Trait ext_websocket::actions::ActionT

pub trait ActionT: Debug {
    fn act_sseq(&self, sseq: &mut Sseq) -> Option<Message> { ... }

    fn act_resolution(
        &self, 
        resolution: &Arc<RwLock<Resolution<CCC>>>
    ) -> Option<Message> { ... }

    fn to_string(&self) -> String { ... }
}

The name Action is sort-of a misnomer. It is the content of any message that is sent between the different objects. There are three functions to implement.

The function user really should be a trait constant, except that makes enum_dispatch unhappy. It indicates whether this action comes from a user action. For example, AddProduct is not a user action but AddProductType is. This field doesn't really make sense for messages sent to the user, but we set it as false anyway (which is the default).

The functions act_sseq and act_resolution executes the action of the command on the corresponding object. These are by default empty and should be left empty if the action is not expected to act on the corresponding object. For example, AddDifferential has an empty act_resolution function.

Provided methods

fn act_sseq(&self, sseq: &mut Sseq) -> Option<Message>

fn act_resolution(
&self,
resolution: &Arc<RwLock<Resolution<CCC>>>
) -> Option<Message>

fn to_string(&self) -> String

Loading content...