Module ext_websocket::sseq

Structs

Differential

Product
Fields

ProductItem

Sseq
Here are some blanket assumptions we make about the order in which we add things.

Enums

ClassState

Constants

INFINITY

MIN_PAGE

_IMPL_DESERIALIZE_FOR_ClassState

_IMPL_DESERIALIZE_FOR_ProductItem

_IMPL_SERIALIZE_FOR_ClassState

_IMPL_SERIALIZE_FOR_ProductItem

Functions

express_basis
Given a vector elt, a subspace zeros of the total space (with a specified choice of complement) and a basis basis of a subspace of the complement, project elt to the complement and express as a linear combination of the basis. This assumes the projection of elt is indeed in the span of basis. The result is returned as a list of coefficients.

sseq_profile

sseq_profile_i