Struct ext_websocket::sseq::Sseq

pub struct Sseq {
    pub p: u32,
    name: SseqChoice,
    min_x: i32,
    min_y: i32,
    pub block_refresh: u32,
    sender: Option<Sender>,
    page_list: Vec<i32>,
    product_name_to_index: HashMap<String, usize>,
    products: Arc<RwLock<Vec<Product>>>,
    classes: BiVec<BiVec<usize>>,
    class_names: BiVec<BiVec<Vec<String>>>,
    differentials: BiVec<BiVec<BiVec<Differential>>>,
    permanent_classes: BiVec<BiVec<Subspace>>,
    zeros: Arc<RwLock<BiVec<BiVec<BiVec<Subspace>>>>>,
    page_classes: Arc<RwLock<BiVec<BiVec<BiVec<(Vec<isize>, Vec<FpVector>)>>>>>,
}

Here are some blanket assumptions we make about the order in which we add things.

If we add a class at (x, y), then all classes to the left and below of (x, y) have been computed. Moreover, every class at (x + 1, y - r) for r >= 1 have been computed. If these have not been set, the class is assumed to be zero.

The same is true for products, where the grading of a product is that of its source.

Whenever a product v . x is set, the target is already set.

Fields

block_refresh : If this is a positive number, then the spectral sequence will not re-compute classes and edges. See Actions::BlockRefresh for details.

Fields
p: u32name: SseqChoicemin_x: i32min_y: i32block_refresh: u32sender: Option<Sender>page_list: Vec<i32>product_name_to_index: HashMap<String, usize>products: Arc<RwLock<Vec<Product>>>classes: BiVec<BiVec<usize>>class_names: BiVec<BiVec<Vec<String>>>differentials: BiVec<BiVec<BiVec<Differential>>>permanent_classes: BiVec<BiVec<Subspace>>zeros: Arc<RwLock<BiVec<BiVec<BiVec<Subspace>>>>>page_classes: Arc<RwLock<BiVec<BiVec<BiVec<(Vec<isize>, Vec<FpVector>)>>>>>
Methods

impl Sseq

pub fn new(
p: u32,
name: SseqChoice,
min_x: i32,
min_y: i32,
sender: Option<Sender>
) -> Self

pub fn clear(&mut self)

This clears out all the user actions. This is intended to be used when we undo, where we clear out all actions then redo the existing actions. Hence we avoid re-allocating as much as possible because we are likely to need the space anyway

pub fn refresh_all(&mut self)

fn add_page(&mut self, r: i32)

Adds a page to the page list, which is the list of pages where something changes from the previous page. This is mainly used by the add_differential function.

fn allocate_differential_matrix(&mut self, r: i32, x: i32, y: i32)

Initializes differentials[x][y][r]. It sets the differentials of all known permament classes to 0.

fn allocate_zeros_subspace(
&self,
zeros: &mut BiVec<BiVec<BiVec<Subspace>>>,
r: i32,
x: i32,
y: i32
)

fn multiply(
&self,
x: i32,
y: i32,
class: &FpVector,
product: &Product
) -> Option<(i32, i32, FpVector)>

Given a class class at (x, y) and a Product object product, compute the product of the class with the product. Returns the new coordinate of the product as well as the actual product. The result is None if the product is not yet computed.

fn leibniz(
&self,
r: i32,
x: i32,
y: i32,
s: &FpVector,
t: Option<&mut FpVector>,
si: usize
) -> Option<(i32, i32, i32, FpVector, Option<FpVector>)>

Apply the Leibniz rule to obtain new differentials. The differential we start with is a d_r differential from (x, y) with source s and target t. If the target is None, then it means s is permanent. In this case, r should be set to INFINITY. On the other hand, if t is zero, it simply means d_r is zero.

The other object we multiply with is the product with index si. If si is permanent, we simply multiply the differenial with si. If si is non-permanent but has a differential starting from it, we apply Leibniz to that differential. If si is the target of a product differential, we do nothing.

If si is permanent and t is None, then we mark s * si to be permanent as well.

There is no good reason why t is a mutable borrow instead of an immutable borrow. It would have been an immutable borrow if we can implicitly cast Option<&mut A> to Option, but we can't.

Return

We return a pair (r_, x_, y_, s_, t_) which is the data of the new differential --- its page, starting coordinate and source and target vectors. Again, if s * si is permanent, then t_ is set to None. If there is not enough products computed to calculate the result, or if s * si is zero, we return None.

fn compute_edges_inner(
x: i32,
y: i32,
p: u32,
name: SseqChoice,
sender: Sender,
page_classes: Arc<RwLock<BiVec<BiVec<BiVec<(Vec<isize>, Vec<FpVector>)>>>>>,
products: Arc<RwLock<Vec<Product>>>,
zeros: Arc<RwLock<BiVec<BiVec<BiVec<Subspace>>>>>
)

fn compute_edges(&self, x: i32, y: i32, parallel: bool)

Computes products whose source is at (x, y).

fn compute_next_page_no_d(
p: u32,
old_classes: &(Vec<isize>, Vec<FpVector>),
zeros: &Subspace
) -> (Vec<isize>, Vec<FpVector>)

Compute the classes in next page assuming there is no differential coming out of the class on that page. Returns a basis of the remaining classes together with column_to_pivot_row.

fn compute_next_page_with_d(
&self,
r: i32,
x: i32,
y: i32,
old_classes: &(Vec<isize>, Vec<FpVector>)
) -> ((Vec<isize>, Vec<FpVector>), Vec<Vec<u32>>)

Compute the classes in next page assuming there might be a differential coming out of the class on that page. Returns a basis of the remaining classes together with column_to_pivot_row.

fn compute_classes(&mut self, x: i32, y: i32, refresh_edge: bool)

Arguments

refresh_edge - Whether to automatically call compute_edges after computing class. This should almost always be yes, unless we are re-computing everything, in which case this will result in many duplicate calls of compute_edge.

fn send_class_data(&self, x: i32, y: i32)

fn send(&self, msg: Message)

impl Sseq

fn product_defined(&self, x: i32, y: i32, product: &Product) -> bool

fn class_defined(&self, x: i32, y: i32) -> bool

fn get_page<T>(r: i32, bivec: &BiVec<T>) -> &T

fn get_differentials_hitting(&self, x: i32, y: i32) -> Vec<i32>

Get a list of r for which there is a d_r differential hitting (x, y)

impl Sseq

pub fn set_class(&mut self, x: i32, y: i32, num: usize)

This function should only be called when everything to the left and bottom of (x, y) has been defined.

pub fn set_class_name(&mut self, x: i32, y: i32, idx: usize, name: String)

pub fn add_differential(
&mut self,
r: i32,
x: i32,
y: i32,
source: &FpVector,
target: &mut FpVector
)

Add a differential starting at (x, y). This mutates the target by reducing it via self.zeros[x - 1][y + r][r]

Panics if the target of the differential is not yet defined

pub fn add_differential_propagate(
&mut self,
r: i32,
x: i32,
y: i32,
source: &FpVector,
target: &mut Option<FpVector>,
product_index: usize
)

This function recursively propagates differentials. If this function is called, it will add the corresponding differential plus all products of index at least product_index. Here we have to exercise a slight bit of care to ensure we don't set both $p_1 p_2 d$ and $p_2 p_1 d$ when $p_1$, $p_2$ are products and $d$ is the differential. Our strategy is that we compute $p_2 p_1 d$ if and only if $p_1$ comes earlier in the list of products than $p_2$.

pub fn add_permanent_class(&mut self, x: i32, y: i32, class: &FpVector)

pub fn add_product_type(
&mut self,
name: &String,
mult_x: i32,
mult_y: i32,
left: bool,
permanent: bool
)

Add a product to the list of products, but don't add any computed product

pub fn add_product_differential(&mut self, source: &String, target: &String)

fn repropagate_product(&mut self, idx: usize)

pub fn add_product(
&mut self,
name: &String,
x: i32,
y: i32,
mult_x: i32,
mult_y: i32,
left: bool,
matrix: &Vec<Vec<u32>>
)